fastapi-auth/requirements.txt
fastapi
SQLAlchemy
Starlette
Pydantic
uvicorn
python-jose
passlib
bcrypt